renamed cuda annotation
